--- notification-chat/src/main/java/com/kor/syh/application/port/in/MessageType.java
@@ -1,5 +1,0 @@
-package com.kor.syh.application.port.in;
-
-public enum MessageType {
-	SEND,SERVER
-}
