+++ notification-chat/src/main/java/com/kor/syh/application/port/in/notification/MessageType.java
@@ -1,0 +1,5 @@
+package com.kor.syh.application.port.in.notification;
+
+public enum MessageType {
+	SEND, SERVER
+}
