+++ notification-chat/src/main/java/com/kor/syh/application/port/in/notification/SendNotificationUseCase.java
@@ -1,4 +1,4 @@
-package com.kor.syh.application.port.in;
+package com.kor.syh.application.port.in.notification;
 
 public interface SendNotificationUseCase {
 	void send(SendMessageCommand sendMessageCommand);
